requirements.txt
# Dev/Deployment
sphinx
sphinx_rtd_theme
nose
coverage
pypi-publisher
kafka-python
requests
